Refactor org creation modal to its own file
